requirements.txt
python-telegram-bot[webhooks]==20.8
yt-dlp
nest_asyncio